refactored constructors for concrete classes inn GameObject heirarchy

+++ src/game_object/GameObject.java
@@ -20,7 +20,7 @@ public abstract class GameObject implements Collider {
 	
 	
 	
-	public GameObject(String imagePath, int sizeWidth, int sizeHeight, Point2D pos) throws FileNotFoundException {
+	public GameObject(String imagePath, double sizeWidth, double sizeHeight, Point2D pos) throws FileNotFoundException {
 		Image image = new Image(new FileInputStream(imagePath));
         myView = new ImageView(image);
