+++ src/game_object/Projectile.java
@@ -9,12 +9,12 @@ public class Projectile extends GameObject implements MoveableTime{
 	protected Point2D myVelocity;
 	protected int myMoveSpeed;
 	
-	public Projectile(String imagePath, int sizeWidth, int sizeHeight, Point2D pos) throws FileNotFoundException {
+	public Projectile(String imagePath, double sizeWidth, double sizeHeight, Point2D pos) throws FileNotFoundException {
 		super( imagePath, sizeWidth, sizeHeight, pos);
 	}
 
 	@Override
-	public void move(double elapsedTime) {
+	public void move(Double elapsedTime) {
 
         myView.setX(myView.getX() + myVelocity.getX() * elapsedTime);
         myView.setY(myView.getY() + myVelocity.getY() * elapsedTime);
